+++ src/matchdom.js
@@ -11,6 +11,7 @@ export * as JsonPlugin from './plugins/json.js';
 export * as NumPlugin from './plugins/number.js';
 export * as DatePlugin from './plugins/date.js';
 export * as DomPlugin from './plugins/dom.js';
+export * as UrlPlugin from './plugins/url.js';
 
 export class Matchdom {
 	static Symbols = Symbols;