+++ CHANGES.md
@@ -131,4 +131,5 @@ Parameters can still be escaped using percent-encoding.
 - core: new mutation type and set:mutation* filter
 - xml, html formats are provided by DomPlugin
 - DomPlugin: `queryAll` is renamed to `all`, `query` is renamed to `one`
+- UrlPlugin: url type, query filter
 - date: add weekday, days, weeks (non-trivial) formats